Добавлен Typescript и мидлвейры
