+++ package.json
@@ -4,11 +4,21 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "test": "echo \"Error no test specified\" && exit 1"
+        "test": "echo \"Error no test specified\" && exit 1",
+        "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+        "build": "tsc",
+        "start": "node dist/index.js"
     },
     "author": "Vladimir",
     "licence": "ISC",
     "dependencies": {
-        "express": "^4.21.2"
+        "express": "^4.21.2",
+        "ts-node-dev": "^2.0.0"
+    },
+    "devDependencies": {
+        "@types/express": "^5.0.0",
+        "@types/node": "^22.13.8",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.8.2"
     }
 }
